Uniformed save and load
Now grids' names are a function of the elements of the system instead of containing 0, 1, 2, as element indicators.
Grids created before this version can be easily imported by swapping the numbers 0 1 2 etc. with the ordered numbers of elements of the potential, both in the grid file names and in the MODULE.json file